recode bash specifics to let sh scripts source freetz_functions
freetz_functions has lots of bash dependent functions,
but most simple functions do not depend on bash.

However, to source these simple freetz_functions from a /bin/sh
source file, we must not use `function name()` but `name()` be-
cause the former is not a valid keyword in /bin/sh.

/bin/sh can now source freetz_functions and will not fail as
long as functions with bash specific code are not called.

This means `echo2` and `sep2` can be sourced to be used in
simple scripts, which can make their output format and
indentation match that of e.g. freetz_patch.