introduce script processing in APPLY_PATCHES (executable *.sh in patch dirs)
Imagine you want to unpack a kernel source and copy files from another
kernel source (possibly unpacking that other source tree only for this
purpose).

You could convert-copy the files to unified diff format, but this would
blur the focus and purpose of patch files.

This patch enables the freetz framework to process shell scripts found
below make/linux/patches - they are processed inline and intermixed
with .patch files found.  Order is important and may be given by select-
ing and prefixing patch/script files with a three or four digit number.

If a <name>.sh and <name>.patch pair exists, such that name is identical
for both suffixes, the patch is processed first and the .sh script is
called thereafter.