optimize zimhttpserver: rework index and search, multiple mainpages
-----

previous indexing method did not distribute article titles
evenly (md5 hash could have been used for that) and for a
small index this wasted diskspace; it also did not allow
non-prefix search, i.e. the term "name" was only looked
for in the set of titles starting with "n"

when serving large zim collections, the indexer in kiwix-
tools may be more suitable, so we can assume the perl script
is used for relatively small title sets only:

thus, simplify the indexing method and write a single file,
using grep as the search engine.  search terms are expected
to be full words, so we use "word boundary" regex feature
to limit search results.  any search will return all the
article titles containing the word searched.  for very
large indexes this may be inefficient, but it performs
well for small and medium sized ZIM files and does not
clutter the file system.

-----

mainpage may vary from zim to zim, try "index" and "mainpage"
variants, with or without html suffix variants